+++ common/assets/styles/etherwallet-master.less
@@ -48,6 +48,4 @@
 @import "etherwallet-swap.less";
 @import "etherwallet-ext-custom.less";
 @import "etherwallet-utilities.less";
-@import "etherwallet-nav.less";
 @import "etherwallet-print.less";
-
