+++ common/assets/styles/etherwallet-utilities.less
@@ -20,11 +20,6 @@
   background: @ether-navy;
 }
 
-.bg-gradient {
-  background: @ether-navy;
-  background: linear-gradient(149deg, #132a45, #143a56, #21a4ce, #19b4ad);
-}
-
 .bg-gray-lighter {
   background-color: @gray-lighter;
 }
