+++ common/components/Footer/index.scss
@@ -1,0 +1,62 @@
+@import "common/sass/variables";
+
+// footer
+.Footer {
+  background-color: $ether-navy;
+  color: white;
+  margin-top: 100px;
+  padding-top: $space-xs;
+  padding-bottom: $space-sm;
+
+  &-logo {
+    width: 100%;
+    height: auto;
+    max-width: 20rem;
+  }
+
+  &-info,
+  &-links {
+    padding-left: 3rem;
+  }
+
+  p {
+    margin: $space-xs 0 $space-sm;
+  }
+
+  a {
+    color: #4ac8ed;
+    &:hover,
+    &:focus {
+      color: darken(#4ac8ed, 5%);
+    }
+  }
+
+  h5 {
+    font-size: $font-size-bump;
+    margin: $font-size-small 0 0;
+
+    i {
+      margin-left: -1.5em;
+      margin-right: .25em;
+    }
+  }
+
+  ul {
+    list-style: none;
+    padding-left: $space-sm;
+    margin: 0 0 $space-xs 0;
+  }
+
+  li,
+  p {
+    font-size: $font-size-small;
+    margin: $space-sm 0;
+  }
+
+  @media screen and (max-width: $grid-float-breakpoint) {
+    .row {
+      margin-left: -.5rem;
+      margin-right: -.5rem;
+    }
+  }
+}
