+++ common/components/Header/components/Navigation.scss
@@ -1,0 +1,116 @@
+@import "common/sass/variables";
+
+.Navigation {
+  position: relative;
+  overflow-y: hidden;
+
+  &-scroll {
+    -ms-overflow-style: -ms-autohiding-scrollbar;
+    -webkit-overflow-scrolling: touch;
+    overflow-x: auto;
+    margin-bottom: -20px;
+    font-size: 0;
+    white-space: nowrap;
+    @media screen and (max-width: $screen-sm) {
+      margin: 0 -5% -20px;
+    }
+  }
+
+  &-links {
+    border-bottom: 2px solid $gray-lighter;
+    display: inline-block;
+    font-size: 0;
+    margin-bottom: 20px;
+    min-width: 100%;
+    padding: 5px 0 0;
+    vertical-align: middle;
+
+    &-item {
+      display: inline-block;
+      font-size: 0;
+
+      &-link {
+        color: darken($link-color, 15%);
+        display: block;
+        font-size: 16px;
+        font-weight: 300;
+        padding: 10px;
+        white-space: nowrap;
+        position: relative;
+        min-height: 2.75rem;
+
+        &:after {
+          content: "";
+          background: $brand-primary;
+          height: 2px;
+          width: 100%;
+          left: 0px;
+          position: absolute;
+          bottom: -1px;
+          transition: all 250ms ease 0s;
+          transform: scaleX(0);
+        }
+
+        &.is-active,
+        &:hover,
+        &:focus {
+          color: $brand-primary;
+          text-decoration: none;
+          transition: all 250ms ease 0s;
+
+          &:after {
+            transform: scaleX(1);
+            transition: all 250ms ease 0s;
+          }
+        }
+      }
+    }
+  }
+
+  &-arrow {
+    background-color: white;
+    bottom: 12px;
+    color: #d6d6d6;
+    font-size: 33px;
+    line-height: 1.3;
+    min-width: 50px;
+    position: absolute;
+    top: 0;
+    vertical-align: middle;
+    width: 5%;
+    z-index: 999;
+    &:hover {
+      text-decoration: none;
+    }
+
+    &--left {
+      left: 3%;
+      background: linear-gradient(
+        to left,
+        rgba(255, 255, 255, 0) 0%,
+        rgba(255, 255, 255, 1) 47%,
+        rgba(255, 255, 255, 1) 100%
+      );
+      padding-left: 5px;
+      text-align: left;
+      @media screen and (max-width: $screen-sm) {
+        left: 0;
+      }
+    }
+
+    &--right {
+      right: 3%;
+      background: linear-gradient(
+        to right,
+        rgba(255, 255, 255, 0) 0%,
+        rgba(255, 255, 255, 1) 47%,
+        rgba(255, 255, 255, 1) 100%
+      );
+      padding-right: 5px;
+      text-align: right;
+      @media screen and (max-width: $screen-sm) {
+        right: 0;
+      }
+    }
+  }
+}
