+++ common/sass/variables.scss
@@ -1,0 +1,281 @@
+@import "./variables/colors";
+@import "./variables/spacing";
+@import "./variables/typography";
+
+// Transitions
+$transition: 500ms all ease-in-out;
+
+// Links
+$link-color: $brand-primary;
+$link-hover-color: darken($link-color, 5%);
+$link-hover-decoration: none;
+
+// Tables
+$table-cell-padding: $space-sm;
+$table-condensed-cell-padding: $space-xs;
+
+$table-bg: transparent;
+$table-bg-accent: #f9f9f9;
+$table-bg-hover: $gray-lightest;
+$table-bg-active: $table-bg-hover;
+
+$table-border-color: #ddd;
+
+// Buttons
+$btn-font-weight: normal;
+
+$btn-default-color: #333;
+$btn-default-bg: #ececec;
+$btn-default-border: $gray-lighter;
+
+$btn-primary-color: #fff;
+$btn-primary-bg: $brand-primary;
+$btn-primary-border: darken($btn-primary-bg, 5%);
+
+$btn-success-color: #fff;
+$btn-success-bg: $brand-success;
+$btn-success-border: darken($btn-success-bg, 5%);
+
+$btn-info-color: #fff;
+$btn-info-bg: $brand-info;
+$btn-info-border: darken($btn-info-bg, 5%);
+
+$btn-warning-color: #fff;
+$btn-warning-bg: $brand-warning;
+$btn-warning-border: darken($btn-warning-bg, 5%);
+
+$btn-danger-color: #fff;
+$btn-danger-bg: $brand-danger;
+$btn-danger-border: darken($btn-danger-bg, 5%);
+
+$btn-link-disabled-color: $gray-light;
+
+// Forms
+$input-bg: #fff;
+$input-bg-disabled: $gray-lightest;
+
+$input-color: $gray;
+$input-border: $gray-lighter;
+
+$input-border-focus: $brand-primary;
+
+$input-color-placeholder: darken($gray-lighter, 10%);
+
+$input-height-base: 2.55rem;
+$input-height-large: 4rem;
+$input-height-small: 2rem;
+
+$form-group-margin-bottom: $space-sm;
+
+$legend-color: $gray-dark;
+$legend-border-color: #e5e5e5;
+
+$input-group-addon-bg: $gray-lighter;
+$input-group-addon-border-color: $input-border;
+
+// Misc.
+$cursor-disabled: default;
+
+$dropdown-bg: #fff;
+$dropdown-border: rgba(0, 0, 0, .15);
+$dropdown-fallback-border: $gray-lighter;
+$dropdown-divider-bg: #e5e5e5;
+
+$dropdown-link-color: $ether-navy;
+$dropdown-link-hover-color: $ether-blue;
+$dropdown-link-hover-bg: $gray-lightest;
+
+$dropdown-link-active-color: $component-active-color;
+$dropdown-link-active-bg: $component-active-bg;
+
+$dropdown-link-disabled-color: $gray-light;
+
+$dropdown-header-color: $gray-light;
+
+$dropdown-caret-color: #000;
+
+$zindex-navbar: 1000;
+$zindex-dropdown: 1000;
+$zindex-popover: 1060;
+$zindex-tooltip: 1070;
+$zindex-navbar-fixed: 1030;
+$zindex-modal-background: 1040;
+$zindex-modal: 1050;
+$zindex-alerts: 1060;
+
+$screen-xs: 32rem;
+$screen-xs-min: $screen-xs;
+
+$screen-sm: 51.2rem;
+$screen-sm-min: $screen-sm;
+
+$screen-md: 66.133333333rem;
+$screen-md-min: $screen-md;
+
+$screen-lg: 80rem;
+$screen-lg-min: $screen-lg;
+
+$screen-xl: 94rem;
+$screen-xl-min: $screen-xl;
+
+$screen-xs-max: ($screen-sm-min - 1);
+$screen-sm-max: ($screen-md-min - 1);
+$screen-md-max: ($screen-lg-min - 1);
+$screen-lg-max: ($screen-xl-min - 1);
+
+$grid-columns: 12;
+$grid-gutter-width: 3rem;
+$grid-float-breakpoint: $screen-sm-min;
+$grid-float-breakpoint-max: ($grid-float-breakpoint - 1);
+
+$cont-padding: 5%;
+$cont-padding-lg: 7.5%;
+
+$container-tablet: ($screen-sm + $grid-gutter-width);
+$container-sm: $container-tablet;
+
+$container-desktop: ($screen-md + $grid-gutter-width);
+$container-md: $container-desktop;
+
+$container-large-desktop: ($screen-lg + $grid-gutter-width);
+$container-lg: $container-large-desktop;
+
+$state-success-text: darken($brand-success, 10%);
+$state-success-bg: #dff0d8;
+$state-success-border: darken(adjust-hue($state-success-bg, -10deg), 5%);
+
+$state-info-text: darken($brand-info, 10%);
+$state-info-bg: #d9edf7;
+$state-info-border: darken(adjust-hue($state-info-bg, -10deg), 7%);
+
+$state-warning-text: darken($brand-warning, 10%);
+$state-warning-bg: #fcf8e3;
+$state-warning-border: darken(adjust-hue($state-warning-bg, -10deg), 5%);
+
+$state-danger-text: darken($brand-danger, 10%);
+$state-danger-bg: #f2dede;
+$state-danger-border: darken(adjust-hue($state-danger-bg, -10deg), 5%);
+
+$tooltip-max-width: 200px;
+$tooltip-color: #fff;
+$tooltip-bg: #000;
+$tooltip-opacity: .9;
+
+$tooltip-arrow-width: $space-sm;
+$tooltip-arrow-color: $tooltip-bg;
+
+$label-default-bg: $gray-light;
+$label-primary-bg: $brand-primary;
+$label-success-bg: $brand-success;
+$label-info-bg: $brand-info;
+$label-warning-bg: $brand-warning;
+$label-danger-bg: $brand-danger;
+
+$label-color: #fff;
+$label-link-hover-color: #fff;
+
+$modal-inner-padding: $space*1.5;
+
+$modal-title-padding: $space;
+$modal-title-line-height: $line-height-base;
+
+$modal-content-bg: #fff;
+$modal-content-border-color: rgba(0, 0, 0, .2);
+$modal-content-fallback-border-color: #999;
+
+$modal-backdrop-bg: #000;
+$modal-backdrop-opacity: .5;
+$modal-header-border-color: #e5e5e5;
+$modal-footer-border-color: $modal-header-border-color;
+
+$modal-lg: 70rem;
+$modal-md: 50rem;
+$modal-sm: 30rem;
+
+$alert-border-radius: $border-radius;
+$alert-link-font-weight: bold;
+
+$alert-success-bg: $brand-success;
+$alert-success-text: white;
+$alert-success-border: $alert-success-bg;
+
+$alert-info-bg: $brand-primary;
+$alert-info-text: white;
+$alert-info-border: $alert-info-bg;
+
+$alert-warning-bg: $brand-warning;
+$alert-warning-text: white;
+$alert-warning-border: $alert-warning-bg;
+
+$alert-danger-bg: $brand-danger;
+$alert-danger-text: white;
+$alert-danger-border: $alert-danger-bg;
+
+$progress-bg: $gray-lightest;
+$progress-bar-color: #fff;
+$progress-border-radius: $border-radius;
+
+$progress-bar-bg: $brand-primary;
+$progress-bar-success-bg: $brand-success;
+$progress-bar-warning-bg: $brand-warning;
+$progress-bar-danger-bg: $brand-danger;
+$progress-bar-info-bg: $brand-info;
+
+$list-group-bg: #fff;
+$list-group-border: #ddd;
+$list-group-border-radius: $border-radius;
+
+$list-group-hover-bg: $gray-lightest;
+$list-group-active-color: $component-active-color;
+$list-group-active-bg: $component-active-bg;
+$list-group-active-border: $list-group-active-bg;
+$list-group-active-text-color: lighten($list-group-active-bg, 40%);
+
+$list-group-disabled-color: $gray-light;
+$list-group-disabled-bg: $gray-lighter;
+$list-group-disabled-text-color: $list-group-disabled-color;
+
+$list-group-link-color: #555;
+$list-group-link-hover-color: $list-group-link-color;
+$list-group-link-heading-color: #333;
+
+$thumbnail-padding: 4px;
+$thumbnail-bg: $body-bg;
+$thumbnail-border: #ddd;
+$thumbnail-border-radius: $border-radius;
+
+$thumbnail-caption-color: $text-color;
+$thumbnail-caption-padding: 9px;
+
+$well-bg: $gray-lightest;
+$well-border: darken($well-bg, 7%);
+
+$badge-font-weight: bold;
+$badge-line-height: 1;
+$badge-border-radius: 10px;
+
+$close-font-weight: bold;
+$close-color: #000;
+$close-text-shadow: 0 1px 0 #fff;
+
+$code-color: #c7254e;
+$code-bg: #f9f2f4;
+
+$kbd-color: #fff;
+$kbd-bg: #333;
+
+$pre-bg: $gray-lightest;
+$pre-color: $gray-dark;
+$pre-border-color: $gray-lighter;
+$pre-scrollable-max-height: 340px;
+
+$component-offset-horizontal: 180px;
+$text-muted: $gray-light;
+$abbr-border-color: $gray-light;
+$headings-small-color: inherit;
+$blockquote-small-color: $gray-light;
+$blockquote-font-size: ($font-size-base * 1.25);
+$blockquote-border-color: $gray-lighter;
+$page-header-border-color: $gray-lighter;
+$dl-horizontal-offset: $component-offset-horizontal;
+$hr-border: $gray-lighter;
