+++ common/sass/variables/typography.scss
@@ -1,0 +1,22 @@
+$font-family-sans-serif: 'Lato', sans-serif;
+$font-family-serif: Georgia, "Times New Roman", Times, serif;
+$font-family-monospace: Menlo, Monaco, Consolas, "Courier New", monospace;
+$font-family-base: $font-family-sans-serif;
+
+$base: 15;
+$font-size-pixels: $base + px;
+$font-size-pixels-xl: $base + 1px; // for xl screens
+$font-size-pixels-sm: $base + px; // for small screens
+
+$font-size-large-bump: 2.25rem; // 33.75
+$font-size-large: 1.90rem; // 28.5
+$font-size-medium-bump: 1.50rem; // 22.5
+$font-size-medium: 1.30rem; // 19.5
+$font-size-bump-more: 1.15rem; // 17.25
+$font-size-bump: 1.07rem; // 16.05
+$font-size-base: 1.00rem; // 15
+$font-size-small: 0.92rem; // 13.8
+$font-size-xs: 0.80rem; // 12
+
+$line-height-base: 1.4;
+$line-height-computed: 1.4;
